Add custom hooks for timer and wake lock management in App component
